pyproject.toml
[project]
name = "pdf-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cloudinary>=1.44.1",
    "python-dotenv>=1.2.1",
    "qrcode>=8.2",
]
